+++ src/lib/form-builder.options.ts
@@ -1,0 +1,5 @@
+import {PreSubmitHandler} from './types';
+
+export class FormBuilderOptions {
+  public static preFormSubmit?: PreSubmitHandler;
+}