+++ src/public-api.ts
@@ -11,6 +11,7 @@ export * from './lib/form-group/form-group.component';
 export * from './lib/form-field/form-field.directive';
 export * from './lib/form-slot/form-slot.directive';
 export * from './lib/form-builder.module';
+export * from './lib/form-builder.options';
 export * from './lib/form-type/abstract-group-type';
 export * from './lib/form-type/abstract-collection-type';
 export * from './lib/form-type/abstract-layout-type';